api-tests/requirements.txt
pytest==7.4.0
requests==2.28.1
pytest-html
python-dotenv
